requirements.txt
numpy
Pillow
fake-useragent
requests
tk
